requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
joblib==1.4.2
numpy==2.2.3
pillow==11.1.0
scikit-learn==1.6.1
python-multipart==0.0.20
